Minor refinements
Everything is ready for everyone to contribute to the functions. I lost
my notes for writing to files, so I don't know how to do that. All of
that would be located in the 'Save' function, Atef said that applying
the string array to a file should be pretty simple, if anyone knows how
to do that, please do.

If anything else is needed from me immediately, let me know.