Decided to add a 'splitToNum' function
Split to num splits the user input from the function that the user is
calling.
I wasn't going to add this, but decided to after I realized that most of
them are the same in terms of user input format.
Also, I made the current line index into a reference variable.