Alterations to move
I changed 'Move' so that it displays the line number as well as prints
the line (it turns out that we're supposed to include something that
displays the line in 'Move').
Using 'Move', I confirmed that the problems lie within 'Insert'.
Also, I'd like to say that I'm running into some issues with the
occasional "(whatever I typed)" is not a valid command.
I'm utterly dumbfounded by this. I have NO IDEA why this is happening.